requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2024.1
Django==5.0.4
django-crispy-forms==2.1
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
idna==3.7
kombu==5.3.7
mysqlclient==2.2.4
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
